Start refactoring the conversion process and add unit test for this process.

+++ app/src/main/java/moe/notify/animenotifier/storage/model/anime/AnimeOpening.java
@@ -12,7 +12,7 @@ import moe.notify.animenotifier.storage.database.AnimeDatabase;
 public class AnimeOpening extends BaseModel {
 
     @Column
-    public float similarity;
+    public double similarity;
     @Column
     public String permalink;
     @Column
